job_bot_project/backend/requirements.txt
fastapi
uvicorn
motor
pydantic[email]
passlib
python-jose
python-dotenv
google-generativeai
PyPDF2
docx
requests
beautifulsoup4
pyotp
APScheduler
slowapi
httpx
pytest-asyncio
PyMuPDF
crewai
google-cloud-speech
